requirements.txt
flask
flask-cors
supabase
bcrypt
requests
python-dotenv
gunicorn
openpyxl
